feat: ...

+++ src/main/java/org/example/zick/domain/user/domain/repository/UserRepository.java
@@ -3,9 +3,10 @@ package org.example.zick.domain.user.domain.repository;
 import org.example.zick.domain.user.domain.User;
 import org.springframework.data.repository.CrudRepository;
 
+import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
     Optional<User> findByUserName(String userName);
     Boolean findAppliedById(Long studentId);
 }
